Update coverage script
